@@ -25,27 +25,32 @@
   "dependencies": {
     "moment": "^2.24.0",
     "moment-range": "^4.0.1",
-    "v-click-outside": "^2.0.2",
-    "vue": "^2.6.9"
+    "vue": "^3.0.2"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.2.0",
-    "@vue/cli-plugin-eslint": "^3.3.0",
-    "@vue/cli-plugin-unit-jest": "^3.5.1",
-    "@vue/cli-service": "^3.2.2",
-    "@vue/eslint-config-standard": "^4.0.0",
-    "@vue/test-utils": "^1.0.0-beta.29",
+    "@vue/cli-plugin-babel": "^4.5.8",
+    "@vue/cli-plugin-eslint": "^4.5.8",
+    "@vue/cli-plugin-unit-jest": "^4.5.8",
+    "@vue/cli-service": "^4.5.8",
+    "@vue/compiler-sfc": "^3.0.2",
+    "@vue/eslint-config-standard": "^5.1.2",
+    "@vue/test-utils": "^2.0.0-beta.7",
     "babel-core": "7.0.0-bridge.0",
-    "babel-eslint": "^10.0.1",
+    "babel-eslint": "^10.1.0",
     "babel-jest": "^23.6.0",
-    "eslint": "^5.8.0",
-    "eslint-plugin-vue": "^5.0.0",
-    "node-sass": "^4.11.0",
+    "core-js": "^3.6.5",
+    "eslint": "^7.0.0-rc.0",
+    "eslint-plugin-import": "^2.20.2",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-standard": "^4.0.0",
+    "eslint-plugin-vue": "^7.0.0-beta.4",
+    "node-sass": "^4.14.1",
     "nuxt": "^2.9.2",
-    "sass-loader": "^7.1.0",
-    "vue-cli-plugin-webpack-bundle-analyzer": "^1.2.0",
-    "vue-server-renderer": "^2.6.9",
-    "vue-template-compiler": "^2.6.9"
+    "sass-loader": "^10.0.4",
+    "typescript": "^4.0.3",
+    "vue-cli-plugin-webpack-bundle-analyzer": "^2.0.0",
+    "vue-jest": "^5.0.0-alpha.5"
   },
   "postcss": {
     "plugins": {
