@@ -7,7 +7,7 @@ describe('VueCtkDateTimePicker/PickersContainer/ButtonValidate', () => {
 
   beforeEach(() => (
     wrapper = shallowMount(ButtonValidate, {
-      propsData: {
+      props: {
         visible: true
       }
     })
@@ -19,30 +19,30 @@ describe('VueCtkDateTimePicker/PickersContainer/ButtonValidate', () => {
     classes.forEach(C => expect(wrapper.classes()).toContain(C))
   })
 
-  it('should have a dark class if the "dark" prop is define', () => {
-    wrapper.setProps({
+  it('should have a dark class if the "dark" prop is define', async () => {
+    await wrapper.setProps({
       dark: true
     })
     expect(wrapper.classes()).toContain('is-dark')
   })
 
-  it('should not have a dark class if the "dark" prop is not defined', () => {
-    wrapper.setProps({
+  it('should not have a dark class if the "dark" prop is not defined', async () => {
+    await wrapper.setProps({
       dark: false
     })
     expect(wrapper.classes()).not.toContain('is-dark')
   })
 
   describe('now button', () => {
-    it('should be defined if not time only, nor the "noButtonNow" prop is defined nor range mode', () => {
-      wrapper.setProps({
+    it('should be defined if not time only, nor the "noButtonNow" prop is defined nor range mode', async () => {
+      await wrapper.setProps({
         noButtonNow: false,
         range: false,
         onlyTime: false
       })
       const button = wrapper.find('.now')
       expect(button.exists()).toBeTruthy()
-      expect(button.is('button')).toBeTruthy()
+      expect(button.element.tagName).toEqual('BUTTON')
       const { tabindex, type } = button.attributes()
       expect(tabindex).toEqual('-1')
       expect(type).toEqual('button')
@@ -55,8 +55,8 @@ describe('VueCtkDateTimePicker/PickersContainer/ButtonValidate', () => {
       expect(buttonContent.text()).toEqual('Now')
     })
 
-    it('should have a custom text if "buttonNowTranslation" prop is defined', () => {
-      wrapper.setProps({
+    it('should have a custom text if "buttonNowTranslation" prop is defined', async () => {
+      await wrapper.setProps({
         noButtonNow: false,
         range: false,
         onlyTime: false,
@@ -66,8 +66,8 @@ describe('VueCtkDateTimePicker/PickersContainer/ButtonValidate', () => {
       expect(buttonContent.text()).toEqual('My button content')
     })
 
-    it('should emit a now event on click', () => {
-      wrapper.setProps({
+    it('should emit a now event on click', async () => {
+      await wrapper.setProps({
         noButtonNow: false,
         range: false,
         onlyTime: false
@@ -77,8 +77,8 @@ describe('VueCtkDateTimePicker/PickersContainer/ButtonValidate', () => {
       expect(wrapper.emitted().now).toBeTruthy()
     })
 
-    it('should be undefined if time only, or the "noButtonNow" prop is defined or range mode', () => {
-      wrapper.setProps({
+    it('should be undefined if time only, or the "noButtonNow" prop is defined or range mode', async () => {
+      await wrapper.setProps({
         noButtonNow: true,
         range: false,
         onlyTime: false
@@ -86,7 +86,7 @@ describe('VueCtkDateTimePicker/PickersContainer/ButtonValidate', () => {
       let button = wrapper.find('.now')
       expect(button.exists()).toBeFalsy()
 
-      wrapper.setProps({
+      await wrapper.setProps({
         noButtonNow: false,
         range: true,
         onlyTime: false
@@ -94,7 +94,7 @@ describe('VueCtkDateTimePicker/PickersContainer/ButtonValidate', () => {
       button = wrapper.find('.now')
       expect(button.exists()).toBeFalsy()
 
-      wrapper.setProps({
+      await wrapper.setProps({
         noButtonNow: false,
         range: false,
         onlyTime: true
@@ -104,8 +104,8 @@ describe('VueCtkDateTimePicker/PickersContainer/ButtonValidate', () => {
     })
 
     describe('right margin', () => {
-      it('should have the class if there is a validate button', () => {
-        wrapper.setProps({
+      it('should have the class if there is a validate button', async () => {
+        await wrapper.setProps({
           noButtonNow: false,
           range: false,
           onlyTime: false,
@@ -115,8 +115,8 @@ describe('VueCtkDateTimePicker/PickersContainer/ButtonValidate', () => {
         expect(button.classes()).toContain('right-margin')
       })
 
-      it('should not have the class if validate button not present', () => {
-        wrapper.setProps({
+      it('should not have the class if validate button not present', async () => {
+        await wrapper.setProps({
           noButtonNow: false,
           range: false,
           onlyTime: false,
@@ -129,21 +129,21 @@ describe('VueCtkDateTimePicker/PickersContainer/ButtonValidate', () => {
   })
 
   describe('validate button', () => {
-    it('should be defined if the "hasButtonValidate" is true', () => {
-      wrapper.setProps({
+    it('should be defined if the "hasButtonValidate" is true', async () => {
+      await wrapper.setProps({
         hasButtonValidate: true
       })
       const button = wrapper.find('.validate')
       expect(button.exists()).toBeTruthy()
-      expect(button.is('button')).toBeTruthy()
+      expect(button.element.tagName).toEqual('BUTTON')
       expect(button.attributes().tabindex).toEqual('-1')
 
       const classes = ['flex', 'align-center', 'justify-content-center']
       classes.forEach(C => expect(button.classes()).toContain(C))
     })
 
-    it('should emit a validate event on click', () => {
-      wrapper.setProps({
+    it('should emit a validate event on click', async () => {
+      await wrapper.setProps({
         hasButtonValidate: true
       })
       const button = wrapper.find('.validate')
@@ -151,8 +151,8 @@ describe('VueCtkDateTimePicker/PickersContainer/ButtonValidate', () => {
       expect(wrapper.emitted().validate).toBeTruthy()
     })
 
-    it('should be undefined if the "hasButtonValidate" is false', () => {
-      wrapper.setProps({
+    it('should be undefined if the "hasButtonValidate" is false', async () => {
+      await wrapper.setProps({
         hasButtonValidate: false
       })
       const button = wrapper.find('.validate')
@@ -166,6 +166,6 @@ describe('VueCtkDateTimePicker/PickersContainer/ButtonValidate', () => {
   })
 
   afterEach(() => {
-    wrapper.destroy()
+    wrapper.unmount()
   })
 })
