@@ -7,7 +7,7 @@ describe('VueCtkDateTimePicker/PickersContainer/TimePicker', () => {
 
   beforeEach(() => (
     wrapper = shallowMount(TimePicker, {
-      propsData: {
+      props: {
         height: 200,
         format: 'HH:mm',
         minuteInterval: 1
@@ -22,15 +22,15 @@ describe('VueCtkDateTimePicker/PickersContainer/TimePicker', () => {
   })
 
   describe('dark class', () => {
-    it('should have the "is-dark" class if the dark prop is specified', () => {
-      wrapper.setProps({
+    it('should have the "is-dark" class if the dark prop is specified', async () => {
+      await wrapper.setProps({
         dark: true
       })
       expect(wrapper.classes()).toContain('is-dark')
     })
 
-    it('should not have the "is-dark" class if the dark prop is not specified', () => {
-      wrapper.setProps({
+    it('should not have the "is-dark" class if the dark prop is not specified', async () => {
+      await wrapper.setProps({
         dark: false
       })
       expect(wrapper.classes()).not.toContain('is-dark')
@@ -38,15 +38,15 @@ describe('VueCtkDateTimePicker/PickersContainer/TimePicker', () => {
   })
 
   describe('inline class', () => {
-    it('should have the "inline" class if the inline prop is specified', () => {
-      wrapper.setProps({
+    it('should have the "inline" class if the inline prop is specified', async () => {
+      await wrapper.setProps({
         inline: true
       })
       expect(wrapper.classes()).toContain('inline')
     })
 
-    it('should not have the "inline" class if the inline prop is not specified', () => {
-      wrapper.setProps({
+    it('should not have the "inline" class if the inline prop is not specified', async () => {
+      await wrapper.setProps({
         inline: false
       })
       expect(wrapper.classes()).not.toContain('inline')
@@ -54,29 +54,27 @@ describe('VueCtkDateTimePicker/PickersContainer/TimePicker', () => {
   })
 
   describe('columns', () => {
-    it('should have 2 columns if there isnt amp', () => {
-      wrapper.setProps({
+    it('should have 2 columns if there isnt amp', async () => {
+      await wrapper.setProps({
         format: 'HH:mm'
       })
 
       const columns = wrapper.findAll('.time-picker-column')
-      expect(columns.exists()).toBeTruthy()
       expect(columns.length).toEqual(2)
 
-      const column = columns.at(0)
+      const column = columns[0]
       const classes = ['flex-1', 'flex', 'flex-direction-column', 'text-center']
       classes.forEach(C => {
         expect(column.classes()).toContain(C)
       })
     })
 
-    it('should have 3 columns if there is apm', () => {
-      wrapper.setProps({
+    it('should have 3 columns if there is apm', async () => {
+      await wrapper.setProps({
         format: 'HH:mm A'
       })
 
       const columns = wrapper.findAll('.time-picker-column')
-      expect(columns.exists()).toBeTruthy()
       expect(columns.length).toEqual(3)
     })
 
@@ -85,8 +83,8 @@ describe('VueCtkDateTimePicker/PickersContainer/TimePicker', () => {
      */
 
     describe('Time periods column', () => {
-      it('should be defined if the format includes time period', () => {
-        wrapper.setProps({
+      it('should be defined if the format includes time period', async () => {
+        await wrapper.setProps({
           format: 'HH:mm A'
         })
 
@@ -94,8 +92,8 @@ describe('VueCtkDateTimePicker/PickersContainer/TimePicker', () => {
         expect(column.exists()).toBeTruthy()
       })
 
-      it('should not be defined if the format does not includes time period', () => {
-        wrapper.setProps({
+      it('should not be defined if the format does not includes time period', async () => {
+        await wrapper.setProps({
           format: 'HH:mm'
         })
 
@@ -104,16 +102,16 @@ describe('VueCtkDateTimePicker/PickersContainer/TimePicker', () => {
       })
 
       describe('items', () => {
-        it('should have 2 items', () => {
-          wrapper.setProps({
+        it('should have 2 items', async () => {
+          await wrapper.setProps({
             format: 'HH:mm A'
           })
 
           const items = wrapper.findAll('.time-picker-column-apms .time-picker-column-item')
-          expect(items.exists()).toBeTruthy()
-          expect(items.is('button')).toBeTruthy()
+          expect(items.length).toBeGreaterThan(0)
+          expect(items.map(item => item.element.tagName)).toEqual(new Array(items.length).fill('BUTTON'))
 
-          const am = items.at(0)
+          const am = items[0]
           expect(am.attributes().type).toEqual('button')
           expect(am.attributes().tabindex).toEqual('-1')
           const classes = ['flex', 'align-center', 'justify-content-center']
@@ -122,7 +120,7 @@ describe('VueCtkDateTimePicker/PickersContainer/TimePicker', () => {
           const amText = am.find('.time-picker-column-item-text')
           expect(amText.text()).toEqual('AM')
 
-          const pm = items.at(1)
+          const pm = items[1]
           const pmText = pm.find('.time-picker-column-item-text')
           expect(pmText.text()).toEqual('PM')
         })
@@ -131,6 +129,6 @@ describe('VueCtkDateTimePicker/PickersContainer/TimePicker', () => {
   })
 
   afterEach(() => {
-    wrapper.destroy()
+    wrapper.unmount()
   })
 })
