@@ -1,4 +1,5 @@
 import { shallowMount } from '@vue/test-utils'
+import { h } from 'vue'
 
 import VueCtkDateTimePicker from '@/VueCtkDateTimePicker'
 import CustomInput from '@/VueCtkDateTimePicker/_subs/CustomInput'
@@ -7,50 +8,48 @@ import PickersContainer from '@/VueCtkDateTimePicker/_subs/PickersContainer'
 describe('VueCtkDateTimePicker', () => {
   let wrapper
 
-  beforeEach(() => (
+  beforeEach(() => {
     wrapper = shallowMount(VueCtkDateTimePicker, {
-      propsData: {}
+      props: {}
     })
-  ))
+  })
 
   describe('input', () => {
     it('should be defined if the "inline" prop is not specified and there isnt a default slot', () => {
       const wrapper = shallowMount(VueCtkDateTimePicker, {
-        propsData: {
+        props: {
           inline: false
         }
       })
 
-      const input = wrapper.find(CustomInput)
+      const input = wrapper.findComponent(CustomInput)
       expect(input.exists()).toBeTruthy()
     })
 
     it('should not be defined if the "inline" prop is specified and there is a slot value', () => {
       const wrapper = shallowMount(VueCtkDateTimePicker, {
-        propsData: {
+        props: {
           inline: true
         },
         slots: {
-          default: '<span>Something</span>'
+          default: h('span', {}, 'Something') // <span>Something</span>
         }
       })
 
-      const input = wrapper.find(CustomInput)
+      const input = wrapper.findComponent(CustomInput)
       expect(input.exists()).toBeFalsy()
     })
 
     describe('id attr', () => {
       it('should have the id prop from the attribute', () => {
         const wrapper = shallowMount(VueCtkDateTimePicker, {
-          propsData: {
-            inline: false
-          },
-          attrs: {
+          props: {
+            inline: false,
             id: 'myField'
           }
         })
 
-        const input = wrapper.find(CustomInput)
+        const input = wrapper.findComponent(CustomInput)
         expect(input.attributes().id).toEqual('myField-input')
       })
     })
@@ -58,15 +57,13 @@ describe('VueCtkDateTimePicker', () => {
     describe('disabled attr', () => {
       it('should have the disabled prop from the attribute', () => {
         const wrapper = shallowMount(VueCtkDateTimePicker, {
-          propsData: {
-            inline: false
-          },
-          attrs: {
+          props: {
+            inline: false,
             disabled: true
           }
         })
 
-        const input = wrapper.find(CustomInput)
+        const input = wrapper.findComponent(CustomInput)
         expect(input.attributes().disabled).toBeDefined()
       })
     })
@@ -74,15 +71,13 @@ describe('VueCtkDateTimePicker', () => {
     describe('name attr', () => {
       it('should have the name attr', () => {
         const wrapper = shallowMount(VueCtkDateTimePicker, {
-          propsData: {
-            inline: false
-          },
-          attrs: {
+          props: {
+            inline: false,
             name: 'myFieldName'
           }
         })
 
-        const input = wrapper.find(CustomInput)
+        const input = wrapper.findComponent(CustomInput)
         expect(input.attributes().name).toEqual('myFieldName')
       })
     })
@@ -91,50 +86,44 @@ describe('VueCtkDateTimePicker', () => {
   describe('picker container', () => {
     it('should be defined if the "disabled" attribute is not defined', () => {
       const wrapper = shallowMount(VueCtkDateTimePicker, {
-        propsData: {
-          inline: false
-        },
-        attrs: {
+        props: {
+          inline: false,
           disabled: false
         }
       })
 
-      const container = wrapper.find(PickersContainer)
+      const container = wrapper.findComponent(PickersContainer)
       expect(container.exists()).toBeTruthy()
     })
 
     it('should not be defined if the "disabled" attribute is defined', () => {
       const wrapper = shallowMount(VueCtkDateTimePicker, {
-        propsData: {
-          inline: false
-        },
-        attrs: {
+        props: {
+          inline: false,
           disabled: true
         }
       })
 
-      const container = wrapper.find(PickersContainer)
+      const container = wrapper.findComponent(PickersContainer)
       expect(container.exists()).toBeFalsy()
     })
 
     describe('id attribute', () => {
       it('should have the id attribute', () => {
         const wrapper = shallowMount(VueCtkDateTimePicker, {
-          propsData: {
-            inline: false
-          },
-          attrs: {
+          props: {
+            inline: false,
             id: 'myField'
           }
         })
 
-        const container = wrapper.find(PickersContainer)
+        const container = wrapper.findComponent(PickersContainer)
         expect(container.attributes().id).toEqual('myField-picker-container')
       })
     })
   })
 
   afterEach(() => {
-    wrapper.destroy()
+    wrapper.unmount()
   })
 })
